feature: fields to change item quantity from the cart

+++ app/src/main/java/alektas/stroymat/data/db/entities/CartItem.java
@@ -31,7 +31,7 @@ public class CartItem extends PricelistItem {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(float quantity) {
         this.quantity = quantity;
     }
 }
